requirements.txt
requests==2.32.3
requests-cache==1.2.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pycountry==24.6.1
beautifulsoup4==4.12.3
lxml==5.2.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
Jinja2==3.1.4
